requirements.txt
Flask
Flask-RESTful
Flask-JWT
Flask-JWT_Extended
Flask-SQLAlchemy
uwsgi
psycopg2
datetime
Pillow
numpy
